move storage client
