+++ pkg/clients/storage.go
@@ -1,4 +1,4 @@
-package storage
+package clients
 
 import (
 	"bytes"
